Merchant rev includes discounts

+++ db/migrate/20210425152728_create_bulk_discounts.rb
@@ -2,7 +2,7 @@ class CreateBulkDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :bulk_discounts do |t|
       t.integer :threshhold
-      t.decimal :discount
+      t.integer :discount
       t.references :merchant,  foreign_key: true
       t.string :name
       t.timestamps
